refactor: move atelet path-input validators to internal/resources
Exported as ValidateActorRef / ValidateAteomUID / ValidateContainerNames /
ValidateRunscHash / ValidateSnapshotURIPrefix so the API server and
controller can apply the same rules at their boundaries, and so the
table tests target a public contract. ValidateContainerNames takes
names rather than the atelet proto to keep resources proto-free.
The RPC-boundary glue (per-request validators) stays in cmd/atelet.
Unifying resources.ValidateActorID with ValidateActorRef is deferred:
it changes ate-api-server's client-facing error responses.
